requirements.txt
torch==2.0.1
transformers==4.30.2
huggingface-hub==0.16.4
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.0
streamlit==1.29.0
faiss-cpu==1.7.4
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3